Change Price data type

+++ src/main/java/com/sreepreetham/product_service/form/ProductRegistrationForm.java
@@ -17,7 +17,7 @@ public class ProductRegistrationForm {
   @NotBlank private String name;
   @NotBlank private String description;
   @NotBlank private BigDecimal price;
-  @Builder.Default private float rating = 0;
+  @Builder.Default private double rating = 0.0;
   @NotBlank private String category;
   @Builder.Default private Boolean inStock = true;
 }
